Stop rewriting history
Now that we’ve started landing and mentioning pull requests, every
rewrite will spam the timeline for those pull requests.

Retconning the history of reworking the pre-book docs helped preserve
the ability to see what changed in migrating to the book, but it was
time-consuming, and soon we will stub our pre-book docs to point to
the book anyway.