Harmonize build instructions with Servo's README
This changes harmonizes the build isntructions in the Servo book with
the ones found in Servo's README. Contents are organized for quickly
finding and jumping to the section you need. Each platform now has its
own page, which allows seperating out platform-specific content such as
troubleshooting tips and gives room to grow new platform-specific
details.

Content about build profiles and mach are combined into the
`building.md` file, but with a suggestion to skip ahead if building
Servo for the first time.

The goal here is to provide a smoother onramp to new developers and to
make it esier to find the content you need when working on a new
platform.

Signed-off-by: Martin Robinson <mrobinson@igalia.com>