Copyedit the Architecture section and combine into a single chapter
This change is a general copy-editing pass on the architecture chapter,
combining the two subchapters into a single one called "Architecture."
Out of date information is updated or removed and the architecture
diagram is also updated to match reality.

Finally, the "Architecture" section of the book is a bit odd having a
chapter called "Architecture" now, so it has been renamed to "Design
Documentation" to make it more general. This section was already holding
things that wouldn't normally be classified as "Architecture" so I think
the new name is a bit more accurate.

Signed-off-by: Martin Robinson <mrobinson@igalia.com>